BOAC-6588: removes aria attribute from highcharts container
